+++ src/onegov.election_day/onegov/election_day/views/vote/main.py
@@ -18,51 +18,7 @@ def view_vote(self, request):
 
     """" The main view. """
 
-    layout = VoteLayout(self, request)
-
-    return {
-        'vote': self,
-        'layout': layout,
-        'show_map': layout.principal.is_year_available(self.date.year)
-    }
-
-
-@ElectionDayApp.html(
-    model=Vote,
-    name='counter-proposal',
-    template='vote/ballot.pt',
-    permission=Public
-)
-def view_vote_counter_proposal(self, request):
-
-    """" The main view (counter-proposal). """
-
-    layout = VoteLayout(self, request, 'counter-proposal')
-
-    return {
-        'vote': self,
-        'layout': layout,
-        'show_map': layout.principal.is_year_available(self.date.year)
-    }
-
-
-@ElectionDayApp.html(
-    model=Vote,
-    name='tie-breaker',
-    template='vote/ballot.pt',
-    permission=Public
-)
-def view_vote_tie_breaker(self, request):
-
-    """" The main view (tie-breaker). """
-
-    layout = VoteLayout(self, request, 'tie-breaker')
-
-    return {
-        'vote': self,
-        'layout': layout,
-        'show_map': layout.principal.is_year_available(self.date.year)
-    }
+    return redirect(VoteLayout(self, request).main_view)
 
 
 @ElectionDayApp.json(