+++ src/onegov.election_day/HISTORY.rst
@@ -1,5 +1,15 @@
 Changelog
 ---------
+
+- Adds vote views for districts.
+  [msom]
+
+- Adds entities and districts heatmaps for candidates and lists.
+  [msom]
+
+- Fixes throwing an error on unexpected principal configuration options.
+  [msom]
+
 3.5.9 (2018-06-19)
 ~~~~~~~~~~~~~~~~~~~